Fix osloader's 64bit compile error about type comparision
If trigger warning_disallowed, avb & slot-ab enable,will find some errors
during osloader's 64bit complie progress.

error:
hardware/intel/kernelflinger/avb/libavb/uefi_avb_ops.c:85:17: error: comparison
between signed and unsigned integer expressions [-Werror=sign-compare]
 if (num_bytes > (partition_size - offset_from_partition))
 ^
 cc1: all warnings being treated as errors

Tracked-On: OAM-57217
Reviewed-on: 618741

+++ avb/libavb/uefi_avb_ops.c
@@ -82,7 +82,7 @@ static AvbIOResult read_from_partition(AvbOps* ops,
   /* Check if num_bytes goes beyond partition end. If so, don't read beyond
    * this boundary -- do a partial I/O instead.
    */
-  if (num_bytes > partition_size - offset_from_partition)
+  if (num_bytes > (size_t)(partition_size - offset_from_partition))
     *out_num_read = partition_size - offset_from_partition;
   else
     *out_num_read = num_bytes;
